1.家庭创建、家庭申请 2.消费分类按家庭隔离

+++ src/main/java/com/xkf/cashbook/mysql/model/IncomeDetailDO.java
@@ -18,7 +18,7 @@ public class IncomeDetailDO {
     @TableId(type = IdType.AUTO)
     private Long id;
 
-    private Integer incomeBy;
+    private Long incomeBy;
 
     private Double incomeAmount;
 
@@ -28,7 +28,7 @@ public class IncomeDetailDO {
 
     private LocalDateTime recordDate;
 
-    private Integer recordBy;
+    private Long recordBy;
 
     private String incomeComment;
